refactor(TimeController): 해당 시간의 예약이 존재하는지 확인 후 삭제
- 존재한다면 예외 발생
- 존재하지 않는다면 정상 삭제

+++ src/main/java/roomescape/controller/api/TimeController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.domain.ReservationTime;
 import roomescape.dto.reservationtime.ReservationTimeRequest;
@@ -20,9 +21,11 @@ import roomescape.dto.reservationtime.ReservationTimeResponse;
 public class TimeController {
 
     private final ReservationTimeDao reservationTimeDao;
+    private final ReservationDao reservationDao;
 
-    public TimeController(ReservationTimeDao reservationTimeDao) {
+    public TimeController(ReservationTimeDao reservationTimeDao, ReservationDao reservationDao) {
         this.reservationTimeDao = reservationTimeDao;
+        this.reservationDao = reservationDao;
     }
 
     @GetMapping("")
@@ -50,6 +53,9 @@ public class TimeController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
+        if (reservationDao.hasReservationOf(id)) {
+            throw new IllegalStateException("해당 시간에 예약이 있어 삭제할 수 없습니다.");
+        }
         reservationTimeDao.deleteReservationTime(id);
 
         return ResponseEntity.ok().build();
